avoid compiler warnings: The field 'RoslynCompletionEngine_old.intellisense' is never used CSharpIntellisense\RoslynCompletionEngine.cs 35 25 CSScriptIntellisense.2017
The field 'SourceCodeFormatter.RoslynFormat' is never used Formatter\SourceCodeFormatter.cs 125 29 CSScriptIntellisense.2017

The primary reference "System.Net.Http", which is a framework assembly, could not be resolved in the currently targeted framework. ".NETFramework,Version=v4.0". To resolve this problem, either remove the reference "System.Net.Http" or retarget your application to a framework version which contains "System.Net.Http".
..\..\Program Files (x86)\Microsoft Visual Studio\2017\Community\MSBuild\15.0\bin\Microsoft.Common.CurrentVersion.targets
2041 5 Intellisense.Common

The primary reference "System.Net.Http", which is a framework assembly, could not be resolved in the currently targeted framework. ".NETFramework,Version=v4.0". To resolve this problem, either remove the reference "System.Net.Http" or retarget your application to a framework version which contains "System.Net.Http".
..\..\Program Files (x86)\Microsoft Visual Studio\2017\Community\MSBuild\15.0\bin\Microsoft.Common.CurrentVersion.targets
2041 5 launcher